Port module so it also works in the browser, update affected code.
