+++ lib/browser/index.js
@@ -15,5 +15,16 @@
  * limitations under the License.
  */
 
-// Note: This file is here for backward-compatibility reasons.
-module.exports = require('./node/graylog');
+var logmagic = require('../common/logmagic')
+var sinks = require('../common/sinks');
+
+(function() {
+  /* Register default browser sinks */
+
+  logmagic.registerSink('console', sinks.console);
+
+  /* Default loggers */
+ logmagic.route('__root__', logmagic.INFO, 'console');
+})();
+
+exports = module.exports = logmagic;
