+++ lib/browser/sinks.js
@@ -15,5 +15,6 @@
  * limitations under the License.
  */
 
-// Note: This file is here for backward-compatibility reasons.
-module.exports = require('./node/graylog');
+var sinks = require('../common/sinks');
+
+module.exports.console = sinks.console;
