+++ lib/common/sinks.js
@@ -15,5 +15,16 @@
  * limitations under the License.
  */
 
-// Note: This file is here for backward-compatibility reasons.
-module.exports = require('./node/graylog');
+module.exports.console = function(modulename, level, message, obj) {
+  var timestamp = (new Date().getTime()) / 1000;
+
+  if (obj && Object.keys(obj).length !== 0) {
+    /* TODO: improve */
+    var fm = obj['full_message'] ? "\n  " + obj['full_message'] : "";
+    obj['full_message'] = undefined;
+    console.log(modulename + " [" + timestamp + "]" + ": "+ message + " " + JSON.stringify(obj) + fm);
+  }
+  else {
+    console.log(modulename + " [" + timestamp + "]" + ": "+ message);
+  }
+};
