+++ lib/node/sinks.js
@@ -15,5 +15,13 @@
  * limitations under the License.
  */
 
-// Note: This file is here for backward-compatibility reasons.
-module.exports = require('./node/graylog');
+var sinks = require('../common/sinks');
+var graylog = require('./graylog');
+
+module.exports.console = sinks.console;
+
+module.exports.gelf = function(modulename, level, message, obj) {
+  /* Outputs a GLEF-style JSON to stderr */
+  var str = graylog.logstr(modulename, level, message, obj);
+  process.stderr.write(str + "\n");
+};
